+++ CHANGELOG.md
@@ -11,7 +11,8 @@
   * Upgraded slingshot to 0.10.2
   * Upgraded data.json to 0.1.2 
   * Added support for Facets
-  * Added support for Contribute
-  * Added support for Flag
+  * Added support for Geopulse
+  * Added support for Reverse Geocoder
+  * Added docs for using World Geographies table
   * Added debug mode